Divided code into header and implementation
